rename chatTextarea to chat
